Delete races (#10) and alert if newer version exists (#11)
Races have always been able to be deleted from the database via SQL, and indeed was exposed
in RaceService. Now, users can right-click on one-to-many selected races, and choose to delete them.

At startup, the app will check GitHub's latest release for handycapper and, if the versions do not match,
will display a dialog notifying the user that a newer version exists. Clicking OK will open a browser tab
to the latest release (on github.com). The app will still start as normal.